requirements.txt
cohere
motor
python-dotenv
ipywidgets
scikit-learn
chromadb
langchain
langchain_community
fastapi
uvicorn
pydantic
pydantic-settings 
python-multipart
pymongo